requirements.txt
aiohappyeyeballs    2.4.3
aiohttp             3.10.10
aiosignal           1.3.1
annotated-types     0.7.0
anyio               4.6.0
async-timeout       4.0.3
attrs               24.2.0
certifi             2024.8.30
cfgv                3.4.0
charset-normalizer  3.4.0
ci-info             0.3.0
click               8.1.7
colorama            0.4.6
configobj           5.0.9
configparser        7.1.0
contourpy           1.3.1
cycler              0.12.1
datasets            3.0.1
dill                0.3.8
distlib             0.3.9
distro              1.9.0
docstring_parser    0.16
etelemetry          0.3.1
eval_type_backport  0.2.0
exceptiongroup      1.2.2
filelock            3.16.1
fire                0.7.0
fitz                0.0.1.dev2
fonttools           4.56.0
frozenlist          1.4.1
fsspec              2024.6.1
groq                0.11.0
h11                 0.14.0
httpcore            1.0.6
httplib2            0.22.0
httpx               0.27.2
huggingface-hub     0.25.2
identify            2.6.1
idna                3.10
importlib_resources 6.4.5
iniconfig           2.0.0
instructor          1.6.3
isodate             0.6.1
Jinja2              3.1.4
jiter               0.5.0
kiwisolver          1.4.8
loguru              0.7.2
looseversion        1.3.0
lxml                5.3.0
markdown-it-py      3.0.0
MarkupSafe          3.0.2
matplotlib          3.10.1
mdurl               0.1.2
multidict           6.1.0
multiprocess        0.70.16
networkx            3.4.1
nibabel             5.3.1
nipype              1.8.6
nodeenv             1.9.1
numpy               2.1.2
openai              1.52.0
packaging           24.1
pandas              2.2.3
pathlib             1.0.1
pillow              10.4.0
pip                 24.2
platformdirs        4.3.6
pluggy              1.5.0
pre_commit          4.0.1
propcache           0.2.0
prov                2.0.1
pyarrow             17.0.0
pydantic            2.9.2
pydantic_core       2.23.4
pydot               3.0.2
pydub               0.25.1
Pygments            2.18.0
PyMuPDF             1.24.11
pymupdf4llm         0.0.17
pyparsing           3.2.0
pypdf               5.0.1
pytest              8.3.3
python-dateutil     2.9.0.post0
python-dotenv       1.0.1
pytz                2024.2
pyxnat              1.6.2
PyYAML              6.0.2
rdflib              6.3.2
requests            2.32.3
rich                13.9.2
scipy               1.14.1
setuptools          75.1.0
shellingham         1.5.4
shortuuid           1.0.13
simplejson          3.19.3
six                 1.16.0
sniffio             1.3.1
swarm               0.1.0
tabulate            0.9.0
tenacity            9.0.0
termcolor           2.5.0
together            1.3.1
tomli               2.0.2
tqdm                4.66.5
traits              6.3.2
typer               0.12.5
typing_extensions   4.12.2
tzdata              2024.2
urllib3             2.2.3
virtualenv          20.27.0
wheel               0.44.0
win32-setctime      1.1.0
xxhash              3.5.0
yarl                1.15.0